#2 Small code restructure.
