+++ frontend/src/main.tsx
@@ -1,7 +1,7 @@
 import { ChakraProvider } from "@chakra-ui/react";
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App";
+import App from "./view/App";
 import { enableMapSet } from "immer";
 
 enableMapSet();
